Introduce a new REST Api framework to communicate with Tuleap
The main goal of this constribution is to make more understable how REST calls are made.
For now, to add 1 REST call , it needs to modify at least 6 classes and interfaces. After the refactoring developers need to modify only 2 classes and interface to make the REST call.

At the end it should be also easier to maintain the plugin and add new REST calls ...

In this patch, the only call implemented is to check if the Tuleap instance is valid or not.

There is no functional change expected here.