[WIP] Add the REST route call with the RESTEAsy framework.
The new REST calls with RESTEasy use the new token authentication when the connection with Tuleap is made.

This contribution also the permit the user to use token authentication along the basic auth. It's always possible to use basic auth for the REST call but it is now deprecated.

REST calls with basic auth will be removed later.